Working version of many to many - pond and ducks

+++ 3._Spring_Data_JPA/ducks/src/main/java/pond/model/DuckRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DuckRepository extends CrudRepository<Duck, Long> {
+public interface DuckRepository extends CrudRepository<Duck, Integer> {
     List<Duck> findByName(String name);
 }
